This PR allows manta to read outer messages from an `io.Reader` instead of a fixed size byte buffer.