@howardchung 2016-07-29
Very excited to see this!  Streaming support was a major reason I picked clarity instead of manta for one of my projects.

I may do some benchmarking now to see how they compare :)