@jcoene 2016-07-29
Implemented a simple reusable stream buffer as per Jordan's suggestion. Starts at 100k and can grow if necessary.

 Before:

```
BenchmarkMatch2159568145-4   	       1	17036238731 ns/op	1203558904 B/op	36311799 allocs/op
BenchmarkReadVarUint32-4     	20000000	       111 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadVarUint64-4     	20000000	       116 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesAligned-4  	100000000	        13.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesUnaligned-4	100000000	        13.3 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/dotabuff/manta	24.744s
```

After:

```BenchmarkMatch2159568145-4   	       1	15669121323 ns/op	1173645496 B/op	35993868 allocs/op
BenchmarkReadVarUint32-4     	20000000	       117 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadVarUint64-4     	20000000	        87.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesAligned-4  	100000000	        14.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesUnaligned-4	100000000	        12.9 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/dotabuff/manta	22.923s
```

TL;DR: small decrease in allocs and bytes - very minor compared to the rest of the parser, but a win's a win.