@jcoene 2016-07-29
Jordan found another unnecessary alloc, small improvement

```
BenchmarkMatch2159568145-4   	       1	15419687553 ns/op	1170640152 B/op	35946692 allocs/op
BenchmarkReadVarUint32-4     	20000000	       121 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadVarUint64-4     	20000000	       104 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesAligned-4  	100000000	        12.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkReadBytesUnaligned-4	100000000	        10.5 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/dotabuff/manta	22.654s
```